+++ packages/core/src/menu/MenuControlState.ts
@@ -1,3 +1,6 @@
-export default interface SuggestControlState<V> {
-  focused: V | undefined
+import {RefObject} from 'react'
+
+export default interface SuggestControlState {
+  refs: RefObject<HTMLDivElement>[]
+  focused: number | undefined
 }