+++ packages/desktop/src/drop-down/DropDown.tsx
@@ -1,10 +1,13 @@
 import React, {FC} from 'react'
 
-import {Overlay, OverlayProps, SimpleTransition, SimpleTransitionProps, css, Pos} from '@qiwi/pijma-core'
+import {Overlay, OverlayProps, SimpleTransition, SimpleTransitionProps, css, Pos, Value} from '@qiwi/pijma-core'
 
 export interface DropDownProps {
   show: boolean
   offset?: number
+  width?: Value
+  minWidth?: Value
+  maxWidth?: Value
   rootClose?: boolean
   target: OverlayProps['target']
   container: OverlayProps['container']
@@ -32,6 +35,9 @@ transition.defaultProps = {
 export const DropDown: FC<DropDownProps> = ({
   show,
   offset,
+  width,
+  minWidth,
+  maxWidth,
   rootClose = true,
   target,
   container,
@@ -47,7 +53,16 @@ export const DropDown: FC<DropDownProps> = ({
     onHide={onHide}
     transition={transition}
     children={(renderProps) => (
-      <Pos type="absolute" minWidth={1} zIndex={999} ref={renderProps.props.ref} mt={offset} css={renderProps.props.style}>
+      <Pos
+        type="absolute"
+        width={width}
+        minWidth={minWidth}
+        maxWidth={maxWidth}
+        zIndex={999}
+        ref={renderProps.props.ref}
+        mt={offset}
+        css={renderProps.props.style}
+      >
         {children}
       </Pos>
     )}